Renamed some files and reorganizing.
