+++ gameCode/game/checkPlacement.js
@@ -68,10 +68,10 @@ export function checkPlacement(placement, board) {
             if(checkColorOfAdjacent(placement, board)) {
                 if(checkNumberOfBug(placement)) {
                     bIsPlacementGood = true
-                } else failureMessage =  "Too many of this bug."
-            } else failureMessage = "Can't place next to a bug of opposite color."
-        } else failureMessage = "Dest Hex not Adjacent to Anything."
-    } else failureMessage = "Hex Occupied."
+                } else failureMessage =  "Too many of this bug." //
+            } else failureMessage = "Can't place next to a bug of opposite color." //
+        } else failureMessage = "Dest Hex not Adjacent to Anything." //
+    } else failureMessage = "Hex Occupied." //
     return [bIsPlacementGood, failureMessage]
 }
