+++ test/fullGameLoop/errorTest.test.js
@@ -1,0 +1,21 @@
+import { expect } from "chai";
+import { gameLoopTestHandle } from "../../gameCode/game/gameLoop.js"
+
+describe("Basic Game Loop Test: Full Game Loop", function (){
+    it("Tests Full Loop: Success, Draw", function (){ 
+
+    })
+
+    it("Tests Errors: Placement Errors", function (){ 
+
+    })
+
+    it("Tests Errors: Movement Errors", function (){ 
+
+    })
+
+    it("Tests Errors: Loop Errors", function (){ 
+
+    })
+
+})
